requirements.txt
flask-sockets
flask
PyJWT